+++ front/src/features/logs/pages/HomePage.tsx
@@ -1,11 +1,9 @@
-import React, { useState } from "react";
-import { useNavigate } from "react-router-dom";
+import React from "react";
 import WelcomeCard from "../components/WelcomeCard";
 import TeamBoard from "../components/TeamBoard";
 import LogGraph from "../components/LogGraph/LogGraph";
 import LogDistribution from "../components/LogDistribution/LogDistribution";
 import RecentLogs from "../components/RecentLogs";
-import SearchBar from "../components/SearchBar";
 import { useQuery } from "@tanstack/react-query";
 import { getLogs } from "../api/logApi";
 import { LogLevel } from "../types/logTypes";
@@ -20,9 +18,6 @@ interface HomePageProps {
 }
 
 const HomePage: React.FC<HomePageProps> = ({ isSidebarOpen }) => {
-  const [searchQuery, setSearchQuery] = useState("");
-  const navigate = useNavigate();
-
   // API로부터 로그 데이터 가져오기
   const { data: logs } = useQuery<ApiLogEntry[]>({
     queryKey: ['logs'],
@@ -33,43 +28,10 @@ const HomePage: React.FC<HomePageProps> = ({ isSidebarOpen }) => {
     retry: false,
   });
 
-  // 검색 실행 함수
-  const handleSearch = () => {
-    if (searchQuery.trim()) {
-      // 검색어를 URL 파라미터로 전달하여 SearchPage로 이동
-      navigate(`/search?query=${encodeURIComponent(searchQuery.trim())}`);
-    }
-  };
-
-  // 엔터 키로 검색
-  const handleKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      handleSearch();
-    }
-  };
-
   return (
     <div className={`flex flex-col ${isSidebarOpen ? 'w-[79.03vw]' : 'w-[87.64vw]'}`}>
-      {/* 서치바와 프로필 */}
-      <div className="flex items-center justify-between w-full">
-        <div className="flex-1">
-          <SearchBar
-            value={searchQuery}
-            onChange={setSearchQuery}
-            placeholder="Search anything"
-            onKeyPress={handleKeyPress}
-          />
-        </div>
-        <div className="flex items-center gap-3 mr-[82px] lg:mr-[58px]">
-          <div className="w-[40px] h-[40px] rounded-full bg-[#496660] flex items-center justify-center text-white font-medium">
-            SY
-          </div>
-          <span className="text-[14px] font-medium text-gray-700">Seyoung</span>
-        </div>
-      </div>
-
       {/* Welcome과 팀보드 영역 */}
-      <div className="flex gap-8 items-stretch w-full mt-8">
+      <div className="flex gap-8 items-stretch w-full pt-3">
         <WelcomeCard className="flex-shrink-0" />
         <div className="flex-1 min-w-0">
           <div className="w-full">
@@ -79,7 +41,7 @@ const HomePage: React.FC<HomePageProps> = ({ isSidebarOpen }) => {
       </div>
 
       {/* 로그 그래프와 파이 차트 */}
-      <section className="flex gap-8 mt-1">
+      <section className="flex gap-8 mt-8">
         <LogGraph projectId={1} />
         <LogDistribution logs={logs} />
       </section>
